refactor(model): rename directory
