+++ components/app/project-detail/ProjectSingleTeamMember.vue
@@ -31,7 +31,7 @@
 </template>
 
 <script>
-import { SocialMedia } from '~/model/SocialMedia';
+import { SocialMedia } from '~/models/SocialMedia';
 
 export default {
   name: 'ProjectSingleTeamMember',
